transaction

+++ database/migrations/2019_07_07_175124_create_purchases_table.php
@@ -21,12 +21,12 @@ class CreatePurchasesTable extends Migration
             $table->integer('store_id')->nullable();
             $table->integer('company_id')->nullable();
             $table->integer('supplier_id')->nullable();
-            $table->integer('discount')->default(0);
+            $table->decimal('discount', 11, 2)->default(0);
             $table->string('discount_string')->nullable();
-            $table->integer('shipping')->default(0);
+            $table->decimal('shipping', 11, 2)->default(0);
             $table->string('shipping_string')->nullable();
-            $table->integer('returns')->default(0);
-            $table->integer('grand_total')->default(0);
+            $table->decimal('returns', 11, 2)->default(0);
+            $table->decimal('grand_total', 11, 2)->default(0);
             $table->integer('credit_days')->nullable();
             $table->date('expiry_date')->nullable();
             $table->string('attachment')->nullable();
